requirements.txt
langchain
langchain-community
langchain-ollama
chromadb
llama-cpp-python
pymupdf           
pdfplumber         # Alternative PDF extraction
python-docx        # DOCX extraction
pytesseract        # OCR for scanned docs
tqdm               # Progress bars
fastapi            # Optional API backend
uvicorn            # ASGI server for FastAPI
requests           # HTTP requests if needed
streamlit          # Web app UI
python-dotenv